requirements.txt
# Plant Disease Detection Chatbot - Dependencies
# Python 3.11+

# ==================== Core Framework ====================
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# ==================== LINE SDK ====================
line-bot-sdk==3.5.1

# ==================== AI/ML ====================
google-generativeai==0.4.0

# ==================== Image Processing ====================
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# ==================== Database ====================
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ==================== Cache ====================
redis[hiredis]==5.0.1

# ==================== HTTP Client ====================
httpx==0.26.0
aiohttp==3.9.3

# ==================== Utilities ====================
python-dotenv==1.0.1
python-multipart==0.0.9

# ==================== Monitoring (Optional) ====================
sentry-sdk[fastapi]==1.40.4

# ==================== Development ====================
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# ==================== Type Stubs ====================
types-redis==4.6.0.20240106
types-Pillow==10.2.0.20240125
